refactor: Remove legacy state.py and fix import paths
- Remove redundant core/state.py (legacy DuckDB-specific implementation)
- Use backend-aware LoadStateManager from state_manager.py instead
- Fix test imports from outdated src.* paths to core.* and vendors.*
- Update documentation to use correct module paths

All tests passing after this change.
